ensure regex specific logic lives in FileDataAsset
